+++ resources/js/components/FormField.vue
@@ -1,24 +1,25 @@
 <template>
     <field-wrapper>
-        <div :class="shownOnDetails?'w-1/4 py-6':'w-1/5 py-6 px-8'" v-if="field.sideLabel">
-            <label class="inline-block text-80 leading-tight">
-                {{ field.name }}
-            </label>
-        </div>
-        <div class="py-6 px-8 w-full" :class="fieldClasses">
-            <div class="flex comodolab-help-field">
-                <div class="pr-4 comodolab-help-field-icon-container" v-html="field.icon" v-if="field.icon"></div>
-                <div>
-                    <h4 v-if="!field.sideLabel && field.name" :class="{'mb-2':field.message}">{{field.name}}</h4>
-                    <div :class="messageClasses">
-                        <div v-if="field.asHtml" v-html="field.message"></div>
-                        <div v-else>{{field.message}}</div>
+        <div class="comodolab-help-field-container w-full">
+            <div :class="shownOnDetails?'w-1/4 py-6':'w-1/5 py-6 px-8'" v-if="field.sideLabel">
+                <label class="inline-block text-80 leading-tight">
+                    {{ field.name }}
+                </label>
+            </div>
+            <div class="py-6 px-8 w-full" :class="fieldClasses">
+                <div class="flex comodolab-help-field">
+                    <div class="pr-4 comodolab-help-field-icon-container" v-html="field.icon" v-if="field.icon"></div>
+                    <div>
+                        <h4 v-if="!field.sideLabel && field.name" :class="{'mb-2':field.message}">{{field.name}}</h4>
+                        <div :class="messageClasses">
+                            <div v-if="field.asHtml" v-html="field.message"></div>
+                            <div v-else>{{field.message}}</div>
+                        </div>
                     </div>
                 </div>
             </div>
         </div>
     </field-wrapper>
-
 </template>
 
 <script>
@@ -59,5 +60,14 @@
 </script>
 
 <style lang="scss">
-
+    .card {
+        > div:nth-child(1) .comodolab-help-field-container {
+            border-radius: .5rem .5rem 0 0;
+            overflow: hidden
+        }
+        > div:last-child .comodolab-help-field-container {
+            border-radius: 0 0 .5rem .5rem ;
+            overflow: hidden
+        }
+    }
 </style>