+++ resources/js/field.js
@@ -1,5 +1,5 @@
 Nova.booting((Vue, router) => {
-    Vue.component('index-help', require('./components/IndexField'));
-    Vue.component('detail-help', require('./components/DetailField'));
-    Vue.component('form-help', require('./components/FormField'));
+    Vue.component('index-help', require('./components/IndexField').default);
+    Vue.component('detail-help', require('./components/DetailField').default);
+    Vue.component('form-help', require('./components/FormField').default);
 })