+++ webpack.mix.js
@@ -1,5 +1,5 @@
 let mix = require('laravel-mix')
 
-mix.setPublicPath('dist')
-   .js('resources/js/field.js', 'js')
-   .sass('resources/sass/field.scss', 'css')
+mix.setPublicPath('./')
+   .js('resources/js/field.js', 'dist/js')
+   .sass('resources/sass/field.scss', 'dist/css')