+++ resources/js/components/FormField.vue
@@ -1,16 +1,20 @@
 <template>
     <field-wrapper>
-        <div class="comodolab-help-field-container w-full">
-            <div :class="shownOnDetails?'w-1/4 py-6':'w-1/5 py-6 px-8'" v-if="field.sideLabel">
+        <div class="comodolab-help-field-container flex w-full">
+            <div :class="labelClasses" v-if="field.sideLabel">
                 <label class="inline-block text-80 leading-tight">
-                    {{ field.name }}
+                    <span v-if="field.asHtml" v-html="field.name"></span>
+                    <span v-else>{{ field.name }}</span>
                 </label>
             </div>
-            <div class="py-6 px-8 w-full" :class="fieldClasses">
+            <div class="px-8" :class="fieldClasses">
                 <div class="flex comodolab-help-field">
                     <div class="pr-4 comodolab-help-field-icon-container" v-html="field.icon" v-if="field.icon"></div>
                     <div>
-                        <h4 v-if="!field.sideLabel && field.name" :class="{'mb-2':field.message}">{{field.name}}</h4>
+                        <template v-if="!field.sideLabel && field.name">
+                            <h4 v-if="field.asHtml" v-html="field.name"></h4>
+                            <h4 v-else :class="{'mb-2':field.message}">{{field.name}}</h4>
+                        </template>
                         <div :class="messageClasses">
                             <div v-if="field.asHtml" v-html="field.message"></div>
                             <div v-else>{{field.message}}</div>
@@ -37,14 +41,36 @@
         },
 
         computed: {
+            labelClasses() {
+                let labelClasses = '';
+
+                labelClasses = this.shownOnDetails ? ' w-1/4 py-4' : ' w-1/5 px-8 py-6';
+
+                if (this.field.fullWidthOnDetail) {
+                    labelClasses += ' px-6';
+                }
+
+                return labelClasses;
+            },
             fieldClasses() {
+                let fieldClasses = '';
+
                 let fieldTypes = {
                     info: 'bg-info text-white',
                     warning: 'bg-warning text-warning-dark',
                     danger: 'bg-danger text-white',
                     header: 'bg-30'
                 };
-                return fieldTypes[this.field.type] || '';
+
+                fieldClasses = fieldTypes[this.field.type] || '';
+
+                if (!this.field.sideLabel) {
+                    fieldClasses += ' w-full';
+                }
+
+                fieldClasses += this.shownOnDetails ? ' w-3/4 py-4' : ' w-4/5 py-6';
+
+                return fieldClasses;
             },
             messageClasses() {
                 let fieldTypes = {
@@ -66,7 +92,7 @@
             overflow: hidden
         }
         > div:last-child .comodolab-help-field-container {
-            border-radius: 0 0 .5rem .5rem ;
+            border-radius: 0 0 .5rem .5rem;
             overflow: hidden
         }
     }