+++ changelog.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Add Laravel 7.0 support (See PR #8 - @weeg)
 
+### Fixed
+- Fix issue with FormField border-radius not being applied for first/last child
+
 ## [0.1.0] - 2019-09-12
 
 ### Changed